UPDATE: add prototype of users, plus - Ws.Tablet.Api

+++ Src/Apps/Tablet/Ws.Tablet.Models/Api/ITabletPluApi.cs
@@ -5,5 +5,5 @@ namespace Ws.Tablet.Models.Api;
 public interface ITabletPluApi
 {
     [Get("/plus")]
-    Task<PluDto> GetPluByNumber(ushort number);
+    Task<PluDto> GetPluByNumber(uint number);
 }
